// book-network-ui/src/app/services/model/bookResponse.ts
/**
 * OpenApi specification - Jayarrakesh
 *
 * Contact: prabakars2003@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookResponse { 
    id?: number;
    title?: string;
    authorname?: string;
    isbn?: string;
    sysnopsis?: string;
    owner?: string;
    cover?: string;
    rate?: number;
    archived?: boolean;
    shareable?: boolean;
}